project/requirements/requirements.txt
# Versioned
git+git://github.com/django/django.git@1.7
django-allauth==0.15.0
djangorestframework
# Temporarily using manual install:
# pip install http://github.com/django-compressor/django-compressor/tarball/develop
django-markdown-deux<1.1

# Unversioned
django-extensions
Pillow
psycopg2
django_compressor

# For django admin-docs
docutils
pygments
